#1701: extract chassis-cluster types into types_chassis.go
First increment of the pkg/config/types.go (2055 LOC) domain split. Moves
the chassis/HA cluster type definitions verbatim out of types.go into a
new same-package sibling file types_chassis.go:

  ChassisConfig, ClusterConfig, RedundancyGroup, InterfaceMonitor,
  IPMonitoring, IPMonitorTarget, EventPolicy, EventWithin,
  BridgeDomainConfig.

Pure byte-identical intra-file motion within package config — every type
stays config.Foo, so the ~194 consumer files that import pkg/config are
untouched and no new package import edge is introduced (no cycle risk).
goimports prunes the now-unused fmt/strconv/strings from the extracted
file. Verified: go build ./... clean, go vet clean, config tests pass.
